doc_pipeline/requirements.txt
python-docx>=1.1.0
python-pptx>=1.0.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
nbformat>=5.9.0
chardet>=5.2.0
